pyproject.toml
[tool.poetry]
name = "product"
version = "0.1.0"
description = ""
authors = ["Hikki25 <adilet.ai@breeze.pro>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
